+++ changelog.md
@@ -3,6 +3,8 @@
 ## 1.1.1
 - [*] Fix bug when coping permissions recursively
 - [+] Use container references for inodeShare instance creation
+- [+] Add default date to share form
+- [+] Add `sharedWith($userId)` method to `InodeQuery`
 ## 1.1.0
 - [+] Added email sharing for files
 - [+] Added user file sharing